[monorepo] Switch to gogoproto code generation, add protoc-gen-gofast to cw-build
This commit switches the protocol buffers generation
to use gogoproto instead of google/protobuf. This should
make the marshalling/unmarshalling steps noticeably faster.
It adds protoc-gen-gofast to the cw-build image and uses
the gofast-out plugin to generate protobuf files.

Right now we're only using the regular file generation
without the gogoproto extensions (which adds vanity functions),
but we could consider this in the future if we need one of these
extensions.

Signed-off-by: AlexandreB <alexandreb@lich.io>